feat(exec): windows exec with runas tool

+++ src/main/java/com/aws/greengrass/util/platforms/windows/WindowsExec.java
@@ -5,29 +5,46 @@
 
 package com.aws.greengrass.util.platforms.windows;
 
+import com.aws.greengrass.config.PlatformResolver;
+import com.aws.greengrass.lifecyclemanager.KernelAlternatives;
 import com.aws.greengrass.util.Exec;
 import com.aws.greengrass.util.Utils;
+import com.aws.greengrass.util.platforms.Platform;
+import com.aws.greengrass.util.platforms.UserPlatform;
 import org.zeroturnaround.process.Processes;
+import vendored.com.microsoft.alm.storage.windows.internal.WindowsCredUtils;
 
 import java.io.File;
 import java.io.IOException;
+import java.nio.ByteBuffer;
+import java.nio.CharBuffer;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 import javax.annotation.Nullable;
+import javax.inject.Inject;
+
+import static com.aws.greengrass.config.PlatformResolver.ARCHITECTURE_KEY;
+import static com.aws.greengrass.config.PlatformResolver.ARCH_X86;
 
 @SuppressWarnings("PMD.AvoidCatchingThrowable")
 public class WindowsExec extends Exec {
-    public static final String PATHEXT_KEY = "PATHEXT";
+    private final List<String> pathext;  // ordered file extensions to try, when no extension is provided
+    private final String runasExePath;
 
-    private static final List<String> PATHEXT;  // ordered file extensions to try, when no extension is provided
-
-    static {
-        String pathExt = System.getenv(PATHEXT_KEY);
-        PATHEXT = Arrays.asList(pathExt.split(File.pathSeparator));
+    @Inject
+    WindowsExec(PlatformResolver platformResolver, KernelAlternatives kernelAlts) {
+        super();
+        String pathExt = System.getenv("PATHEXT");
+        pathext = Arrays.asList(pathExt.split(File.pathSeparator));
+        Path runasPath = ARCH_X86.equals(platformResolver.getCurrentPlatform().get(ARCHITECTURE_KEY))
+                ? kernelAlts.getBinDir().resolve("runas_x86.exe") : kernelAlts.getBinDir().resolve("runas_x64.exe");
+        runasExePath = runasPath.toAbsolutePath().toString();
     }
 
     @Nullable
@@ -40,7 +57,7 @@ public class WindowsExec extends Exec {
                 return Files.isExecutable(f) ? f : null;
             }
             // No extension provided. Try PATHEXT in order
-            for (String extCandidate : PATHEXT) {
+            for (String extCandidate : pathext) {
                 Path f = Paths.get(fn + extCandidate);
                 if (Files.isExecutable(f)) {
                     return f;
@@ -55,7 +72,7 @@ public class WindowsExec extends Exec {
                 return f;
             }
             // No extension provided. Try PATHEXT in order
-            for (String extCandidate : PATHEXT) {
+            for (String extCandidate : pathext) {
                 f = d.resolve(fn + extCandidate);
                 if (Files.isExecutable(f)) {
                     return f;
@@ -67,25 +84,46 @@ public class WindowsExec extends Exec {
 
     @Override
     public String[] getCommand() {
-        String[] decorated = cmds;
+        String[] decorated = Arrays.copyOf(cmds, cmds.length);
         if (shellDecorator != null) {
             decorated = shellDecorator.decorate(decorated);
-        }
-        // First item in the command is the executable. If it's given as absolute path, add quotes around it
-        // in case the path contains space which will break the whole command line
-        // See security remarks:
-        // https://docs.microsoft.com/en-us/windows/win32/api/winbase/nf-winbase-createprocesswithlogonw
-        if (isAbsolutePath(decorated[0])) {
-            decorated[0] = String.format("\"%s\"", decorated[0]);
         }
         return decorated;
     }
 
     @Override
     protected Process createProcess() throws IOException {
+        if (needToSwitchUser()) {
+            return createRunasProcess();
+        } else {
+            ProcessBuilder pb = new ProcessBuilder();
+            pb.environment().putAll(environment);
+            return pb.directory(dir).command(getCommand()).start();
+        }
+    }
+
+    private Process createRunasProcess() throws IOException {
+        String username = userDecorator.getUser();
+
+        byte[] credBlob = WindowsCredUtils.read(username);
+        ByteBuffer bb = ByteBuffer.wrap(credBlob);
+        CharBuffer cb = StandardCharsets.UTF_8.decode(bb);
+
+        List<String> args = new ArrayList<>();
+        args.add(runasExePath);
+        args.add("-u:" + username);
+        args.add("-p:" + cb);
+        args.add("-l:off");  // disable logging
+        args.addAll(Arrays.asList(getCommand()));
+
+        Arrays.fill(cb.array(), (char) 0);  // zero-out temporary buffers
+        Arrays.fill(bb.array(), (byte) 0);
+
         ProcessBuilder pb = new ProcessBuilder();
         pb.environment().putAll(environment);
-        return pb.directory(dir).command(getCommand()).start();
+        Process p = pb.directory(dir).command(args).start();
+        args.clear();  // best effort to clear password presence
+        return p;
     }
 
     @Override
@@ -99,14 +137,29 @@ public class WindowsExec extends Exec {
         Process killerProcess = new ProcessBuilder().command("taskkill", "/f", "/t", "/pid",
                 Integer.toString(Processes.newPidProcess(process).getPid())).start();
         try {
-            killerProcess.waitFor();
-            process.destroyForcibly();
-            process.waitFor(5, TimeUnit.SECONDS);
+            int taskkillExitCode = killerProcess.waitFor();
+            if (taskkillExitCode != 0) {
+                process.destroyForcibly();
+                process.waitFor(5, TimeUnit.SECONDS);
+            }
         } catch (InterruptedException e) {
             Thread.currentThread().interrupt();
         } finally {
             process.destroyForcibly();
         }
+    }
+
+    /**
+     * Returns true if we need to create process as another user. Otherwise, just use ProcessBuilder.
+     */
+    private boolean needToSwitchUser() throws IOException {
+        if (userDecorator == null) {
+            return false;
+        }
+        // check if same as current user
+        UserPlatform.UserAttributes currUser = Platform.getInstance().lookupCurrentUser();
+        return !(currUser.getPrincipalName().equals(userDecorator.getUser()) || currUser.getPrincipalIdentifier()
+                .equals(userDecorator.getUser()));
     }
 
     private static boolean isAbsolutePath(String p) {
